Appended BUILD_ID to the name field
This change updates the name field passed to the Bitbucket API to reflect the intended use of these fields. The key field should be just the name of the job, where the name field should be the name of the job plus the build number, "An identifier for the build itself, e.g. BB-DEPLOY-1".