Fix Bitbucket cloud bugs and add Paths feature
When using bitbucket cloud the direction source parameters was making
the resulting version to not always use the correct branch, which would
then fail in the `get` step because the commit was not found in the
branch.

The feature to add exclude_title was not adding the variable to the jq
call.

Added ability to filter PR based on paths, fixes #58